Extract fragment visual reconciliation
